reorganise tally files.
